remove versions and other fixes
Warn user that if removing object using versions, it do not create a
delete marker.
So if not using versions i do not pass versions_id to action.

Delete marker option show delete marker ONLY.
If i show both, i have two problems: there is no way to identify a
delete marker obj (i can identify it as it do not have a Size key), but
main problem is that they are put all at the end of the other one, so i
no more have a proper key order.

Use print in place of logger, this way can be parsed using other prog
(Ex grep).